sonar fixes

+++ hoot-core/src/main/cpp/hoot/core/io/HootApiDbWriter.cpp
@@ -376,40 +376,40 @@ void HootApiDbWriter::writeChange(const Change& change)
 
 void HootApiDbWriter::_createElement(const ConstElementPtr& element)
 {
-  switch (element->getElementType().getEnum())
+  if (element->getElementType() == ElementType::Node)
   {
-    case ElementType::Node:
-      _hootdb.insertNode(std::dynamic_pointer_cast<const Node>(element));
-      break;
-    //only supporting nodes for now
-    default:
-      throw HootException("Unsupported element type");
+    _hootdb.insertNode(std::dynamic_pointer_cast<const Node>(element));
+  }
+  else
+  {
+    // only supporting nodes for now
+    throw HootException("Unsupported element type");
   }
 }
 
 void HootApiDbWriter::_modifyElement(const ConstElementPtr& element)
 {
-  switch (element->getElementType().getEnum())
+  if (element->getElementType() == ElementType::Node)
   {
-    case ElementType::Node:
-      _hootdb.updateNode(std::dynamic_pointer_cast<const Node>(element));
-      break;
-    //only supporting nodes for now
-    default:
-      throw HootException("Unsupported element type");
+    _hootdb.updateNode(std::dynamic_pointer_cast<const Node>(element));
+  }
+  else
+  {
+    // only supporting nodes for now
+    throw HootException("Unsupported element type");
   }
 }
 
 void HootApiDbWriter::_deleteElement(const ConstElementPtr& element)
 {
-  switch (element->getElementType().getEnum())
+  if (element->getElementType() == ElementType::Node)
   {
-    case ElementType::Node:
-      _hootdb.deleteNode(std::dynamic_pointer_cast<const Node>(element));
-      break;
-    //only supporting nodes for now
-    default:
-      throw HootException("Unsupported element type");
+    _hootdb.deleteNode(std::dynamic_pointer_cast<const Node>(element));
+  }
+  else
+  {
+    // only supporting nodes for now
+    throw HootException("Unsupported element type");
   }
 }
